Remove redundant TestRedshiftQueryFailureRecoverySmokeTest
